refactor user's types
