+++ packages/client/src/Redux/user/types.ts
@@ -1,4 +1,15 @@
-import { TProfileInfo } from '../../types'
+export type TProfileInfo = {
+  first_name: string
+  second_name: string
+  phone: string
+  email: string
+  display_name: string
+  login: string
+  password: string
+  id: number
+  avatar: string
+  role?: string
+}
 
 export type TUserState = TProfileInfo | null
